This PR fixes two issues:
- Local transactions should not be subject to the `future shouldn't churn pending txs` rule
- Transactions are added to the queue first, and on the next reorg procedure promoted 

There's still an issue with this, as the rule can be gamed if the queue contains a real future transaction with nonce -1.