@rjl493456442 2023-03-17
```diff
diff --git a/core/txpool/list.go b/core/txpool/list.go
index 724bb6caca..adba934d69 100644
--- a/core/txpool/list.go
+++ b/core/txpool/list.go
@@ -272,8 +272,8 @@ func newList(strict bool) *list {
 
 // Overlaps returns whether the transaction specified has the same nonce as one
 // already contained within the list.
-func (l *list) Overlaps(tx *types.Transaction) bool {
-	return l.txs.Get(tx.Nonce()) != nil
+func (l *list) Overlaps(nonce uint64) bool {
+	return l.txs.Get(nonce) != nil
 }
 
 // Add tries to insert a new transaction into the list, returning whether the
diff --git a/core/txpool/txpool.go b/core/txpool/txpool.go
index ad556f39fb..465e5c27a4 100644
--- a/core/txpool/txpool.go
+++ b/core/txpool/txpool.go
@@ -741,7 +741,7 @@ func (pool *TxPool) add(tx *types.Transaction, local bool) (replaced bool, err e
 			var replacesPending bool
 			for _, dropTx := range drop {
 				dropSender, _ := types.Sender(pool.signer, dropTx)
-				if list := pool.pending[dropSender]; list != nil && list.Overlaps(dropTx) {
+				if list := pool.pending[dropSender]; list != nil && list.Overlaps(dropTx.Nonce()) {
 					replacesPending = true
 					break
 				}
@@ -766,7 +766,7 @@ func (pool *TxPool) add(tx *types.Transaction, local bool) (replaced bool, err e
 	}
 
 	// Try to replace an existing transaction in the pending pool
-	if list := pool.pending[from]; list != nil && list.Overlaps(tx) {
+	if list := pool.pending[from]; list != nil && list.Overlaps(tx.Nonce()) {
 		// Nonce already pending, check if required price bump is met
 		inserted, old := list.Add(tx, pool.config.PriceBump)
 		if !inserted {
@@ -811,16 +811,24 @@ func (pool *TxPool) add(tx *types.Transaction, local bool) (replaced bool, err e
 
 // isFuture reports whether the given transaction is immediately executable.
 func (pool *TxPool) isFuture(from common.Address, tx *types.Transaction) bool {
-	list := pool.pending[from]
-	if list == nil {
-		return pool.pendingNonces.get(from) != tx.Nonce()
+	// Short circuit if transaction matches pending nonce and can be promoted
+	// to pending list as as executable one.
+	next := pool.pendingNonces.get(from)
+	if tx.Nonce() == next {
+		return false
 	}
-	// Sender has pending transactions.
-	if old := list.txs.Get(tx.Nonce()); old != nil {
-		return false // It replaces a pending transaction.
+	// The transaction has a nonce gap with pending list, it's only considered
+	// as executable if transactions in queue can fill up the nonce gap.
+	queue, ok := pool.queue[from]
+	if !ok {
+		return true
 	}
-	// Not replacing, check if parent nonce exists in pending.
-	return list.txs.Get(tx.Nonce()-1) == nil
+	for nonce := next; nonce < tx.Nonce(); nonce++ {
+		if !queue.Overlaps(nonce) {
+			return true // txs in queue can't fill up the nonce gap
+		}
+	}
+	return false
 }
 
 // enqueueTx inserts a new transaction into the non-executable transaction queue.

```

It's an alternative solution for checking if the transaction is executable quickly.