@MariusVanDerWijden 2023-03-17
@rjl493456442 I've applied your suggestion, sounds reasonable to me. The node only needs to go through the whole queued list if the queued list only contains pending transactions. This should be very rare as those pending transactions in queued are promoted on every reorg.