@MariusVanDerWijden 2023-03-29
master:
```
BenchmarkListAdd-8                     	       5	 213115027 ns/op	55201067 B/op	 2899749 allocs/op
BenchmarkPendingDemotion100-8          	 8279304	       142.7 ns/op	       0 B/op	       0 allocs/op
BenchmarkPendingDemotion1000-8         	 9195732	       132.0 ns/op	       0 B/op	       0 allocs/op
BenchmarkPendingDemotion10000-8        	 8333773	       162.6 ns/op	       0 B/op	       0 allocs/op
BenchmarkFuturePromotion100-8          	503466724	         2.685 ns/op	       0 B/op	       0 allocs/op
BenchmarkFuturePromotion1000-8         	447070483	         2.472 ns/op	       0 B/op	       0 allocs/op
BenchmarkFuturePromotion10000-8        	485711005	         2.447 ns/op	       0 B/op	       0 allocs/op
BenchmarkBatchInsert100-8              	     100	  10090007 ns/op	  365838 B/op	    8576 allocs/op
BenchmarkBatchInsert1000-8             	      12	  94658675 ns/op	 3674252 B/op	   84852 allocs/op
BenchmarkBatchInsert10000-8            	       2	 857325474 ns/op	27979196 B/op	  628158 allocs/op
BenchmarkBatchLocalInsert100-8         	     100	  10363587 ns/op	 1020245 B/op	    9219 allocs/op
BenchmarkBatchLocalInsert1000-8        	      14	  91235529 ns/op	22965913 B/op	   91097 allocs/op
BenchmarkBatchLocalInsert10000-8       	       1	1148621300 ns/op	98535400 B/op	  823841 allocs/op
BenchmarkInsertRemoteWithAllLocals-8   	      39	  29979346 ns/op	 3667078 B/op	   78090 allocs/op
BenchmarkMultiAccountBatchInsert-8     	    9150	    336670 ns/op	   19867 B/op	     109 allocs/op
PASS
```
This PR:
```
BenchmarkListAdd-8                     	       5	 204985908 ns/op	55201048 B/op	 2899749 allocs/op
BenchmarkPendingDemotion100-8          	 9353367	       125.1 ns/op	       0 B/op	       0 allocs/op
BenchmarkPendingDemotion1000-8         	 9497776	       132.0 ns/op	       0 B/op	       0 allocs/op
BenchmarkPendingDemotion10000-8        	 9449646	       126.1 ns/op	       0 B/op	       0 allocs/op
BenchmarkFuturePromotion100-8          	527678170	         2.277 ns/op	       0 B/op	       0 allocs/op
BenchmarkFuturePromotion1000-8         	521783406	         2.325 ns/op	       0 B/op	       0 allocs/op
BenchmarkFuturePromotion10000-8        	517455364	         2.290 ns/op	       0 B/op	       0 allocs/op
BenchmarkBatchInsert100-8              	     158	   7955420 ns/op	  360536 B/op	    8340 allocs/op
BenchmarkBatchInsert1000-8             	      14	  75580800 ns/op	 3644127 B/op	   83885 allocs/op
BenchmarkBatchInsert10000-8            	       2	 735908672 ns/op	27980436 B/op	  628145 allocs/op
BenchmarkBatchLocalInsert100-8         	     151	   8529386 ns/op	 2690471 B/op	    9252 allocs/op
BenchmarkBatchLocalInsert1000-8        	      15	  82062286 ns/op	26405937 B/op	   91173 allocs/op
BenchmarkBatchLocalInsert10000-8       	       2	 912423954 ns/op	443202844 B/op	  872722 allocs/op
BenchmarkInsertRemoteWithAllLocals-8   	      36	  29946683 ns/op	 3666904 B/op	   78180 allocs/op
BenchmarkMultiAccountBatchInsert-8     	    9615	    265377 ns/op	   20614 B/op	     107 allocs/op
```