@holiman 2023-03-29
Re the benchmarks, I tested on master with the following change: 
```
// isFuture reports whether the given transaction is immediately executable.
func (pool *TxPool) isFuture(from common.Address, tx *types.Transaction) bool {
	panic(1)

```
All benchmarks passed -- so none of them actually cover the code that this PR touches.