@MariusVanDerWijden 2023-03-29
Master: 
```
BenchmarkFutureAttack-8   	    6468	    164469 ns/op	    7983 B/op	     184 allocs/op
```
PR:
```
BenchmarkFutureAttack-8   	    7389	    163814 ns/op	    7987 B/op	     184 allocs/op
```

Sorry about that, added a benchmark that hits the spot (it hits the average case, not the worst case though)